+++ e2e/cypress/integration/content-search.js
@@ -11,7 +11,6 @@ LOGIN_TYPES.forEach((loginType) => {
       cy.clearCookies()
       cy.login(loginType)
 
-      cy.interceptGetObject('@canido')
       cy.interceptGetObject('@addable-types')
 
       cy.get(ITEMS_PANELS_SELECTORS.table)
@@ -24,36 +23,26 @@ LOGIN_TYPES.forEach((loginType) => {
       cy.addContent('second folder', 'second-folder', 'btnAddFolder')
       cy.addContent('first item', 'first-item', 'btnAddItem')
       cy.addContent('second item', 'second-item', 'btnAddItem')
-      cy.addGMI('Test GMI item', 'test-gmi-item')
+      cy.addGMI({
+        name: 'Test GMI item',
+        id: 'test-gmi-item',
+      })
 
       cy.get(CONTEXT_TOOLBAR_SELECTORS.selectFilteType).select('Item')
-      cy.get(ITEMS_PANELS_SELECTORS.table)
-        .find('tbody')
-        .find('tr')
-        .its('length')
-        .should('eq', 2)
+      cy.totalItemsInTable(ITEMS_PANELS_SELECTORS.table, 2)
 
       cy.get(CONTEXT_TOOLBAR_SELECTORS.selectFilteType).select('Folder')
-      cy.get(ITEMS_PANELS_SELECTORS.table)
-        .find('tbody')
-        .find('tr')
-        .its('length')
-        .should('eq', 2)
+      cy.totalItemsInTable(ITEMS_PANELS_SELECTORS.table, 2)
 
       cy.get(CONTEXT_TOOLBAR_SELECTORS.selectFilteType).select('GMI')
-      cy.get(ITEMS_PANELS_SELECTORS.table)
-        .find('tbody')
-        .find('tr')
-        .its('length')
-        .should('eq', 1)
+      cy.totalItemsInTable(ITEMS_PANELS_SELECTORS.table, 1)
     })
 
     it('Search items in first level', () => {
       cy.interceptGetObject('test-folder')
-      cy.interceptGetObject('@search')
+      cy.interceptGetObject('@search**')
 
       cy.wait('@get-object-@addable-types')
-      cy.wait('@get-object-@canido')
 
       cy.addContent('First item', 'first-item', 'btnAddItem')
       cy.addContent('Second Item', 'second-item', 'btnAddItem')
@@ -75,15 +64,51 @@ LOGIN_TYPES.forEach((loginType) => {
         )}/`
       )
       cy.wait('@get-object-@addable-types')
-      cy.wait('@get-object-@canido')
 
       cy.get(CONTEXT_TOOLBAR_SELECTORS.selectFilteType).select('Item')
-      cy.wait('@get-object-@search')
-      cy.get(ITEMS_PANELS_SELECTORS.table)
-        .find('tbody')
-        .find('tr')
-        .its('length')
-        .should('eq', 2)
+      cy.wait('@get-object-@search**')
+      cy.totalItemsInTable(ITEMS_PANELS_SELECTORS.table, 2)
+    })
+
+    it('Creates 20 GMI and use the filters', () => {
+      cy.findByText('GMI Folder').click()
+      cy.get(
+        `[data-test='${ITEMS_PANELS_SELECTORS.prefixFilterItem}choice_field']`
+      ).select('keyword')
+      cy.totalItemsInTable(ITEMS_PANELS_SELECTORS.table, 8)
+
+      cy.get(
+        `[data-test='${ITEMS_PANELS_SELECTORS.prefixFilterItem}choice_field']`
+      ).select('date')
+
+      cy.totalItemsInTable(ITEMS_PANELS_SELECTORS.table, 9)
+      cy.get(`[data-test='tag-date']`).find('button').click()
+      cy.totalItemsInTable(ITEMS_PANELS_SELECTORS.table, 10)
+      cy.get(
+        `[data-test='${ITEMS_PANELS_SELECTORS.prefixFilterItem}boolean_field']`
+      ).select('false')
+      cy.findAllByText('25 items', { exact: false })
+      cy.get(`[data-test='tag-false']`).find('button').click()
+      cy.findAllByText('50 items', { exact: false })
+      cy.get(
+        `[data-test='${ITEMS_PANELS_SELECTORS.prefixFilterItem}number_field']`
+      )
+        .clear()
+        .type('19')
+      cy.totalItemsInTable(ITEMS_PANELS_SELECTORS.table, 1)
+    })
+
+    it('Sort by name', () => {
+      cy.findByText('GMI Folder').click()
+      cy.findByText('Test GMI item 0').should('not.exist')
+      cy.get(
+        `[data-test='${ITEMS_PANELS_SELECTORS.prefixSortableItem}title']`
+      ).click()
+      cy.findByText('Test GMI item 9').should('exist')
+      cy.get(
+        `[data-test='${ITEMS_PANELS_SELECTORS.prefixSortableItem}title']`
+      ).click()
+      cy.findByText('Test GMI item 0').should('exist')
     })
   })
 })