+++ e2e/example/pages/_app.js
@@ -1,6 +1,0 @@
-import React from 'react'
-import '../node_modules/react-gmi/dist/css/style.css'
-
-export default function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} />
-}