+++ guillotina_example/guillotina_react_app/guillotina_react_app/api.py
@@ -1,0 +1,9 @@
+from guillotina import configure
+
+
+@configure.service(method='POST', name='@foobar',
+                   permission='guillotina.AccessContent')
+async def example_service(context, request):
+    return {
+        'foo': 'bar'
+    }